+++ tests/spec/responses/schema.ts
@@ -82,7 +82,7 @@ export type RequestParams = Omit<RequestInit, "body" | "method"> & {
   secure?: boolean;
 };
 
-export type RequestQueryParamsType = Record<string, string | string[] | number | number[] | boolean | undefined>;
+export type RequestQueryParamsType = Record<string | number, any>;
 
 type ApiConfig<SecurityDataType> = {
   baseUrl?: string;
@@ -100,6 +100,10 @@ type TPromise<ResolveType, RejectType = any> = Omit<Promise<ResolveType>, "then"
     onrejected?: ((reason: RejectType) => TResult | PromiseLike<TResult>) | undefined | null,
   ): TPromise<ResolveType | TResult, RejectType>;
 };
+
+const enum BodyType {
+  Json,
+}
 
 class HttpClient<SecurityDataType> {
   public baseUrl: string = "https://6-dot-authentiqio.appspot.com/";
@@ -127,17 +131,27 @@ class HttpClient<SecurityDataType> {
 
   private addQueryParam(query: RequestQueryParamsType, key: string) {
     return (
-      encodeURIComponent(key) +
-      "=" +
-      encodeURIComponent(Array.isArray(query[key]) ? (query[key] as any).join(",") : query[key])
+      encodeURIComponent(key) + "=" + encodeURIComponent(Array.isArray(query[key]) ? query[key].join(",") : query[key])
     );
   }
 
-  protected addQueryParams(query?: RequestQueryParamsType): string {
-    const fixedQuery = query || {};
-    const keys = Object.keys(fixedQuery).filter((key) => "undefined" !== typeof fixedQuery[key]);
-    return keys.length === 0 ? "" : `?${keys.map((key) => this.addQueryParam(fixedQuery, key)).join("&")}`;
+  protected addQueryParams(rawQuery?: RequestQueryParamsType): string {
+    const query = rawQuery || {};
+    const keys = Object.keys(query).filter((key) => "undefined" !== typeof query[key]);
+    return keys.length
+      ? `?${keys
+          .map((key) =>
+            typeof query[key] === "object" && !Array.isArray(query[key])
+              ? this.addQueryParams(query[key] as object).substring(1)
+              : this.addQueryParam(query, key),
+          )
+          .join("&")}`
+      : "";
   }
+
+  private bodyFormatters: Record<BodyType, (input: any) => any> = {
+    [BodyType.Json]: JSON.stringify,
+  };
 
   private mergeRequestOptions(params: RequestParams, securityParams?: RequestParams): RequestParams {
     return {
@@ -163,13 +177,14 @@ class HttpClient<SecurityDataType> {
     method: string,
     { secure, ...params }: RequestParams = {},
     body?: any,
+    bodyType?: BodyType,
     secureByDefault?: boolean,
   ): TPromise<T, E> =>
     fetch(`${this.baseUrl}${path}`, {
       // @ts-ignore
       ...this.mergeRequestOptions(params, (secureByDefault || secure) && this.securityWorker(this.securityData)),
       method,
-      body: body ? JSON.stringify(body) : null,
+      body: body ? this.bodyFormatters[bodyType || BodyType.Json](body) : null,
     }).then(async (response) => {
       const data = await this.safeParseResponse<T, E>(response);
       if (!response.ok) throw data;
@@ -197,7 +212,7 @@ export class Api<SecurityDataType = any> extends HttpClient<SecurityDataType> {
      * @response `default` `Error`
      */
     keyRevokeNosecret: (query: { email: string; phone: string; code?: string }, params?: RequestParams) =>
-      this.request<{ status?: string }, Error>(`/key${this.addQueryParams(query)}`, "DELETE", params, null),
+      this.request<{ status?: string }, Error>(`/key${this.addQueryParams(query)}`, "DELETE", params),
 
     /**
      * @tags key, post
@@ -222,7 +237,7 @@ export class Api<SecurityDataType = any> extends HttpClient<SecurityDataType> {
      * @response `default` `Error`
      */
     keyRevoke: (PK: string, query: { secret: string }, params?: RequestParams) =>
-      this.request<{ status?: string }, Error>(`/key/${PK}${this.addQueryParams(query)}`, "DELETE", params, null),
+      this.request<{ status?: string }, Error>(`/key/${PK}${this.addQueryParams(query)}`, "DELETE", params),
 
     /**
      * @tags key, get
@@ -235,7 +250,7 @@ export class Api<SecurityDataType = any> extends HttpClient<SecurityDataType> {
      * @response `default` `Error`
      */
     getKey: (PK: string, params?: RequestParams) =>
-      this.request<{ since?: string; status?: string; sub?: string }, Error>(`/key/${PK}`, "GET", params, null),
+      this.request<{ since?: string; status?: string; sub?: string }, Error>(`/key/${PK}`, "GET", params),
 
     /**
      * @tags key, head
@@ -247,7 +262,7 @@ export class Api<SecurityDataType = any> extends HttpClient<SecurityDataType> {
      * @response `410` `Error` Key is revoked `revoked-key`
      * @response `default` `Error`
      */
-    headKey: (PK: string, params?: RequestParams) => this.request<any, Error>(`/key/${PK}`, "HEAD", params, null),
+    headKey: (PK: string, params?: RequestParams) => this.request<any, Error>(`/key/${PK}`, "HEAD", params),
 
     /**
      * @tags key, post
@@ -315,7 +330,7 @@ export class Api<SecurityDataType = any> extends HttpClient<SecurityDataType> {
      * @response `default` `Error`
      */
     signDelete: (job: string, params?: RequestParams) =>
-      this.request<{ status?: string }, Error>(`/scope/${job}`, "DELETE", params, null),
+      this.request<{ status?: string }, Error>(`/scope/${job}`, "DELETE", params),
 
     /**
      * @tags scope, get
@@ -328,7 +343,7 @@ export class Api<SecurityDataType = any> extends HttpClient<SecurityDataType> {
      * @response `default` `Error`
      */
     signRetrieve: (job: string, params?: RequestParams) =>
-      this.request<{ exp?: number; field?: string; sub?: string }, Error>(`/scope/${job}`, "GET", params, null),
+      this.request<{ exp?: number; field?: string; sub?: string }, Error>(`/scope/${job}`, "GET", params),
 
     /**
      * @tags scope, head
@@ -341,7 +356,7 @@ export class Api<SecurityDataType = any> extends HttpClient<SecurityDataType> {
      * @response `default` `Error`
      */
     signRetrieveHead: (job: string, params?: RequestParams) =>
-      this.request<any, Error>(`/scope/${job}`, "HEAD", params, null),
+      this.request<any, Error>(`/scope/${job}`, "HEAD", params),
 
     /**
      * @tags scope, post
@@ -355,7 +370,7 @@ export class Api<SecurityDataType = any> extends HttpClient<SecurityDataType> {
      * @response `default` `Error`
      */
     signConfirm: (job: string, params?: RequestParams) =>
-      this.request<{ status?: string }, Error>(`/scope/${job}`, "POST", params, null),
+      this.request<{ status?: string }, Error>(`/scope/${job}`, "POST", params),
 
     /**
      * @tags scope, put
@@ -368,6 +383,6 @@ export class Api<SecurityDataType = any> extends HttpClient<SecurityDataType> {
      * @response `default` `Error`
      */
     signUpdate: (job: string, params?: RequestParams) =>
-      this.request<{ jwt?: string; status?: string }, Error>(`/scope/${job}`, "PUT", params, null),
+      this.request<{ jwt?: string; status?: string }, Error>(`/scope/${job}`, "PUT", params),
   };
 }